Remove unimplemented gaiacli init command
