+++ docs/examples/democoin/cmd/democli/main.go
@@ -55,7 +55,6 @@ func main() {
 	// add standard rpc, and tx commands
 
 	rootCmd.AddCommand(
-		rpc.InitClientCommand(),
 		rpc.StatusCommand(),
 		client.LineBreak,
 		tx.SearchTxCmd(cdc),
